python_docs/ds_docs/requirements.txt
blinker==1.8.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
gensim==4.3.2
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.0
nltk==3.8.1
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.15
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
six==1.16.0
smart-open==7.0.4
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.4
tzdata==2024.1
Werkzeug==3.0.3
wordcloud==1.9.3
wrapt==1.16.0
